move project to use cmake
