+++ .gitignore
@@ -1,3 +1,4 @@
+build
 src/docconf
 .project
 .cproject
