+++ CMakeLists.txt
@@ -1,0 +1,8 @@
+cmake_minimum_required(VERSION 3.6)
+
+set(CMAKE_DISABLE_IN_SOURCE_BUILD ON)
+
+enable_testing()
+
+add_subdirectory(src)
+add_subdirectory(test)
