--- Makefile.am
@@ -1,2 +1,0 @@
-SUBDIRS = src test
-ACLOCAL_AMFLAGS = -I m4
