--- autogen.sh
@@ -1,1 +1,0 @@
-autoreconf --force --install
