--- configure.ac
@@ -1,30 +1,0 @@
-#                                               -*- Autoconf -*-
-# Process this file with autoconf to produce a configure script.
-
-AC_PREREQ([2.68])
-AC_INIT([libcollectc], [0.0.0], [srdjan.panic@ymail.com])
-AC_CONFIG_SRCDIR([src/array.c])
-AC_CONFIG_HEADERS([config.h])
-
-#Setup automake
-AM_INIT_AUTOMAKE([foreign subdir-objects])
-
-#Libtool init
-LT_INIT([shared]) 
-
-#Libtool macro dir
-AC_CONFIG_MACRO_DIR([m4])
-
-# Checks for programs.
-AC_PROG_CC
-
-# Checks for libraries.
-
-# Checks for header files.
-
-# Checks for typedefs, structures, and compiler characteristics.
-
-# Checks for library functions.
-
-AC_CONFIG_FILES([Makefile src/Makefile test/Makefile])
-AC_OUTPUT
