+++ src/include/array.h
