+++ test/CMakeLists.txt
@@ -1,0 +1,56 @@
+cmake_minimum_required(VERSION 3.6)
+project(collectc_test)
+
+set(sources test.c)
+
+set(array_test_sources array_test.c)
+set(deque_test_sources deque_test.c)
+set(list_test_sources list_test.c)
+set(hashset_test_sources hashset_test.c)
+set(hashtable_test_sources hashtable_test.c)
+set(queue_test_sources queue_test.c)
+set(slist_test_sources slist_test.c)
+set(stack_test_sources stack_test.c)
+set(test_test_sources test.c)
+set(treeset_test_sources treeset_test.c)
+set(treetable_test_sources treetable_test.c)
+
+include_directories(${PROJECT_SOURCE_DIR}/include ${collectc_INCLUDE_DIRS})
+
+add_executable(array_test ${array_test_sources})
+add_executable(deque_test ${deque_test_sources})
+add_executable(hashset_test ${hashset_test_sources})
+add_executable(hashtable_test ${hashtable_test_sources})
+add_executable(list_test ${list_test_sources})
+add_executable(queue_test ${queue_test_sources})
+add_executable(slist_test ${slist_test_sources})
+add_executable(stack_test ${stack_test_sources})
+add_executable(treeset_test ${treeset_test_sources})
+add_executable(treetable_test ${treetable_test_sources})
+
+add_library(${PROJECT_NAME} SHARED ${sources})
+
+set_target_properties(${PROJECT_NAME} PROPERTIES POSITION_INDEPENDENT_CODE ON)
+target_link_libraries(${PROJECT_NAME})
+
+target_link_libraries(array_test collectc ${PROJECT_NAME})
+target_link_libraries(deque_test collectc ${PROJECT_NAME})
+target_link_libraries(list_test collectc ${PROJECT_NAME})
+target_link_libraries(hashset_test collectc ${PROJECT_NAME})
+target_link_libraries(hashtable_test collectc ${PROJECT_NAME})
+target_link_libraries(queue_test collectc ${PROJECT_NAME})
+target_link_libraries(slist_test collectc ${PROJECT_NAME})
+target_link_libraries(stack_test collectc ${PROJECT_NAME})
+target_link_libraries(treeset_test collectc ${PROJECT_NAME})
+target_link_libraries(treetable_test collectc ${PROJECT_NAME})
+
+add_test(ArrayTest array_test)
+add_test(DequeTest deque_test)
+add_test(ListTest list_test)
+add_test(HashSetTest hashset_test)
+add_test(HashTableTest hashtable_test)
+add_test(QueueTest queue_test)
+add_test(SlistTest slist_test)
+add_test(StackTest stack_test)
+add_test(TreeSetTest treeset_test)
+add_test(TreeTableTest treetable_test)
